refactor: 플레이어 카드 컴포넌트 분리
- 게임 페이지 전체 재랜더링 방지
